layout bootstrao design system
Full-width layout that spans the entire browser window width on all screen sizes. Contains 7 elements. On smaller screens they stack in a 1x7 grid, on medium screens they form groups of 1 on top and two on the bottom, and on large screens there is a single large element on the top and the rest form a 3x2 grid