Getting to Know U(RL)
Create a simple, multi-page website as a way of introducing yourself to the class. The entire site should have a consistent navigation interface that is usable across all of the pages (so that the user can navigate from any page to any other page without ever using the Internet browser “back” button). The site should be structured so that there is one subpage for each of the following prompts